src/api/requirements.txt
azure-identity
openai
Quart
